[frontend] move components into modules folder
