+++ apps/frontend/src/pages/user/reset.jsx
@@ -1,38 +1,5 @@
-import { sendPasswordReset } from "src/firebase"
-import { Button, Flex, Center, Heading, Input } from "@chakra-ui/react"
-import Link from "next/link"
-import { useState } from "react"
+import ResetForm from "src/components/modules/user/resetform"
 
-const Reset = () => {
-  const [email, setEmail] = useState("")
+const Reset = () => <ResetForm />
 
-  return (
-    <Center h="100vh" w="100vw">
-      <Flex direction="column" bg="gray.100" p={12} rounded="md">
-        <Heading mb={6} align="center">
-          Password Reset
-        </Heading>
-        <Input
-          placeholder="email"
-          variant="filled"
-          mb={3}
-          value={email}
-          onChange={(e) => setEmail(e.target.value)}
-        />
-        <Button
-          colorScheme="teal"
-          mb={6}
-          onClick={() => sendPasswordReset(email)}
-        >
-          Send Reset Link to Email
-        </Button>
-        <Link href="/user/signup" passHref>
-          <Button variant="link" size="sm">
-            Sign Up
-          </Button>
-        </Link>
-      </Flex>
-    </Center>
-  )
-}
 export default Reset
